Fix getting AP key, add type functions, update logic and tests
